requirements.txt
discord.py
python-dotenv
sqlalchemy
apscheduler